+++ bin/makefile-dos
@@ -13,12 +13,12 @@ LDFLAGS = -g graphics.lib binmode.lib mouse.lib
 
 RM = del
 
-SRCFILES = conspage.c gcoflow.c shift.c dbgtool.c gcr.c         gcrcell.c llstk.c gcscan.c loopsops.c   storage.c allocmds.c dir.c gvar2.c lowlev1.c    subr.c arith2.c lowlev2.c subr0374.c    arith3.c doscomm.c hardrtn.c lsthandl.c sxhash.c        arith4.c draw.c main.c testtool.c array.c dsk.c         inet.c misc7.c timer.c array2.c dspif.c initdsp.c       miscn.c typeof.c array3.c initkbd.c ubf1.c      array4.c dspsubrs.c initsout.c mkatom.c ubf2.c  array5.c eqf.c intcall.c mkcell.c ubf3.c array6.c       ether.c mkvdate.c ufn.c atom.c findkey.c        kbdsubrs.c mouseif.c ufs.c bbtsub.c foreign.c   keyevent.c unixcomm.c bin.c fp.c binds.c       asmbbt.c fvar.c mvs.c unwind.c bitblt.c gc.c    uraid.c blt.c gc2.c kprint.c osmsg.c usrsubr.c byteswap.c gcarray.c  perrno.c ldeboot.c ldeether.c uutils.c carcdr.c gccode.c        rawcolor.c vars3.c gcfinal.c ldsout.c return.c  vmemsave.c chardev.c gchtfind.c lineblt8.c rpc.c        xc.c common.c gcmain3.c lisp2c.c rplcons.c z2.c xmkicon.c xbbt.c xinit.c xscroll.c xcursor.c xlspwin.c  xrdopt.c xwinman.c      dosmouse.c vesafns.asm vesainit.c vgainit.c kbdif.c     dspsparc.il copyright launch.asm lpread.c lpsolve.c lpmain.c lpwrite.c lpdual.c lptran.c
+SRCFILES = conspage.c gcoflow.c shift.c dbgtool.c gcr.c         gcrcell.c llstk.c gcscan.c loopsops.c   storage.c allocmds.c dir.c gvar2.c lowlev1.c    subr.c arithops.c lowlev2.c subr0374.c doscomm.c hardrtn.c lsthandl.c sxhash.c draw.c main.c testtool.c array.c dsk.c         inet.c misc7.c timer.c array2.c dspif.c initdsp.c       miscn.c typeof.c array3.c initkbd.c ubf1.c      array4.c dspsubrs.c initsout.c mkatom.c ubf2.c  array5.c eqf.c intcall.c mkcell.c ubf3.c array6.c       ether.c mkvdate.c ufn.c atom.c findkey.c        kbdsubrs.c mouseif.c ufs.c bbtsub.c foreign.c   keyevent.c unixcomm.c bin.c fp.c binds.c       asmbbt.c fvar.c mvs.c unwind.c bitblt.c gc.c    uraid.c blt.c gc2.c kprint.c osmsg.c usrsubr.c byteswap.c gcarray.c  perrno.c ldeboot.c ldeether.c uutils.c carcdr.c gccode.c        rawcolor.c vars3.c gcfinal.c ldsout.c return.c  vmemsave.c chardev.c gchtfind.c lineblt8.c rpc.c        xc.c common.c gcmain3.c lisp2c.c rplcons.c z2.c xmkicon.c xbbt.c xinit.c xscroll.c xcursor.c xlspwin.c  xrdopt.c xwinman.c      dosmouse.c vesafns.asm vesainit.c vgainit.c kbdif.c     dspsparc.il copyright launch.asm lpread.c lpsolve.c lpmain.c lpwrite.c lpdual.c lptran.c
 
-OFILES = conspage.obj gcoflow.obj shift.obj dbgtool.obj         gcr.obj  gcrcell.obj llstk.obj  gcscan.obj loopsops.obj storage.obj     allocmds.obj dir.obj gvar2.obj lowlev1.obj      subr.obj arith2.obj lowlev2.obj       subr0374.obj arith3.obj doscomm.obj     hardrtn.obj lsthandl.obj sxhash.obj arith4.obj  draw.obj main.obj testtool.obj array.obj        dsk.obj inet.obj misc7.obj timer.obj    array2.obj dspif.obj initdsp.obj miscn.obj      typeof.obj array3.obj initkbd.obj ubf1.obj      array4.obj dspsubrs.obj initsout.obj    mkatom.obj ubf2.obj array5.obj eqf.obj  intcall.obj mkcell.obj ubf3.obj array6.obj      ether.obj ufn.obj atom.obj      findkey.obj kbdsubrs.obj mouseif.obj ufs.obj    bbtsub.obj foreign.obj keyevent.obj     unixcomm.obj bin.obj fp.obj binds.obj fvar.obj mvs.obj      unwind.obj bitblt.obj gc.obj    uraid.obj blt.obj gc2.obj       kprint.obj osmsg.obj usrsubr.obj byteswap.obj   gcarray.obj perrno.obj uutils.obj       carcdr.obj asmbbt.obj gccode.obj        vars3.obj gcfinal.obj ldsout.obj        return.obj vmemsave.obj chardev.obj     gchtfind.obj lineblt8.obj rpc.obj xc.obj        common.obj gcmain3.obj lisp2c.obj rplcons.obj   z2.obj  vdate.obj $(COLORFILES) $(ARCHFILES) $(LPFILES)
+OFILES = conspage.obj gcoflow.obj shift.obj dbgtool.obj         gcr.obj  gcrcell.obj llstk.obj  gcscan.obj loopsops.obj storage.obj     allocmds.obj dir.obj gvar2.obj lowlev1.obj      subr.obj arithops.obj lowlev2.obj       subr0374.obj doscomm.obj     hardrtn.obj lsthandl.obj sxhash.obj draw.obj main.obj testtool.obj array.obj        dsk.obj inet.obj misc7.obj timer.obj    array2.obj dspif.obj initdsp.obj miscn.obj      typeof.obj array3.obj initkbd.obj ubf1.obj      array4.obj dspsubrs.obj initsout.obj    mkatom.obj ubf2.obj array5.obj eqf.obj  intcall.obj mkcell.obj ubf3.obj array6.obj      ether.obj ufn.obj atom.obj      findkey.obj kbdsubrs.obj mouseif.obj ufs.obj    bbtsub.obj foreign.obj keyevent.obj     unixcomm.obj bin.obj fp.obj binds.obj fvar.obj mvs.obj      unwind.obj bitblt.obj gc.obj    uraid.obj blt.obj gc2.obj       kprint.obj osmsg.obj usrsubr.obj byteswap.obj   gcarray.obj perrno.obj uutils.obj       carcdr.obj asmbbt.obj gccode.obj        vars3.obj gcfinal.obj ldsout.obj        return.obj vmemsave.obj chardev.obj     gchtfind.obj lineblt8.obj rpc.obj xc.obj        common.obj gcmain3.obj lisp2c.obj rplcons.obj   z2.obj  vdate.obj $(COLORFILES) $(ARCHFILES) $(LPFILES)
 
 
-HFILES = address.h adr68k.h arith.h cell.h dbprint.h display.h  dspif.h ifpage.h iopage.h lispemul.h lispmap.h  lsptypes.h miscstat.h lspglob.h array.h bb.h    bitblt.h debug.h devconf.h dspdata.h ether.h    fast_dsp.h gcdata.h hdw_conf.h initatms.h inlinec.h keyboard.h         lispver1.h lispver2.h lldsp.h locfile.h medleyfp.h mouseif.h my.h  opcodes.h osmsg.h pilotbbt.h print.h return.h stack.h stream.h subrs.h timeout.h   tos1defs.h tosfns.h tosret.h vmemsave.h         xdefs.h xbitmaps.h xkeymap.h
+HFILES = address.h adr68k.h arithopsdefs.h arith.h cell.h dbprint.h display.h  dspif.h ifpage.h iopage.h lispemul.h lispmap.h  lsptypes.h miscstat.h lspglob.h array.h bb.h    bitblt.h debug.h devconf.h dspdata.h ether.h    fast_dsp.h gcdata.h hdw_conf.h initatms.h inlinec.h keyboard.h         lispver1.h lispver2.h lldsp.h locfile.h medleyfp.h mouseif.h my.h  opcodes.h osmsg.h pilotbbt.h print.h return.h stack.h stream.h subrs.h timeout.h   tos1defs.h tosfns.h tosret.h vmemsave.h         xdefs.h xbitmaps.h xkeymap.h
 
 
 
@@ -202,9 +202,9 @@ subr.obj : ../src/subr.c
 	$(CC) @copts ../src/subr.c -I ../inc -c $@
 	del copts
 
-arith2.obj : ../src/arith2.c
+arithops.obj : ../src/arithops.c
 	@ echo  $(CFLAGS) > copts
-	$(CC) @copts ../src/arith2.c -I ../inc -c $@
+	$(CC) @copts ../src/arithops.c -I ../inc -c $@
 	del copts
 
 lowlev2.obj : ../src/lowlev2.c
@@ -215,11 +215,6 @@ lowlev2.obj : ../src/lowlev2.c
 subr0374.obj : ../src/subr0374.c
 	@ echo  $(CFLAGS) > copts
 	$(CC) @copts ../src/subr0374.c -I ../inc -c $@
-	del copts
-
-arith3.obj : ../src/arith3.c
-	@ echo  $(CFLAGS) > copts
-	$(CC) @copts ../src/arith3.c -I ../inc -c $@
 	del copts
 
 doscomm.obj : ../src/doscomm.c
@@ -240,11 +235,6 @@ lsthandl.obj : ../src/lsthandl.c
 sxhash.obj : ../src/sxhash.c
 	@ echo  $(CFLAGS) > copts
 	$(CC) @copts ../src/sxhash.c -I ../inc -c $@
-	del copts
-
-arith4.obj : ../src/arith4.c
-	@ echo  $(CFLAGS) > copts
-	$(CC) @copts ../src/arith4.c -I ../inc -c $@
 	del copts
 
 draw.obj : ../src/draw.c
