--- inc/arith4defs.h
@@ -1,9 +1,0 @@
-#ifndef ARITH4DEFS_H
-#define ARITH4DEFS_H 1
-#include "lispemul.h" /* for LispPTR */
-LispPTR N_OP_times2(int tosm1, int tos);
-LispPTR N_OP_itimes2(int tosm1, int tos);
-LispPTR N_OP_quot(int tosm1, int tos);
-LispPTR N_OP_iquot(int tosm1, int tos);
-LispPTR N_OP_iremainder(int tosm1, int tos);
-#endif
